%in%,stars-method {stars}	R Documentation

evaluate whether cube values are in a given set

Description

evaluate whether cube values are in a given set

Usage

## S4 method for signature 'stars'
x %in% table

Arguments

x
data cube value

table
values of the set

[Package stars version 0.7-1 Index]