read ncdf file into stars object

Read data from a file using the NetCDF library directly.

read_ncdf(.x, ..., var = NULL, ncsub = NULL,
  curvilinear = character(0))

Arguments

.x
NetCDF file or source

...
ignored

var
variable name or names (they must be on matching grids)

ncsub
matrix of start, count columns

curvilinear
length two character vector with names of subdatasets holding longitude and latitude values for all raster cells.

Details

The following logic is applied to coordinate axes. If any coordinate axes have regularly spaced coordinates they are reduced to the offset/delta form with 'affine = c(0, 0)', otherwise the values of the coordinates are stored. If the data has two or more dimensions and the first two are regular they are nominated as the 'raster' for plotting.

If `var` is not set the first set of variables on a shared grid is used. It's supposed to be the grid with the most dimensions, but there's no control yet (see `ncmeta::nc_grids(.x)` for the current assumption).

start and count columns of ncsub must correspond to the variable dimemsion (nrows) and be valid index using `ncdf4::ncvar_get` convention (start is 1-based).

Examples

f <- system.file("nc/reduced.nc", package = "stars")
read_ncdf(f)

#> stars object with 4 dimensions and 4 attributes
#> attribute(s):
#>      anom               err             ice             sst       
#>  Min.   :-10.160   Min.   :0.110   Min.   :0.010   Min.   :-1.80  
#>  1st Qu.: -0.580   1st Qu.:0.160   1st Qu.:0.470   1st Qu.:-0.03  
#>  Median : -0.080   Median :0.270   Median :0.920   Median :13.65  
#>  Mean   : -0.186   Mean   :0.263   Mean   :0.718   Mean   :12.99  
#>  3rd Qu.:  0.210   3rd Qu.:0.320   3rd Qu.:0.960   3rd Qu.:24.81  
#>  Max.   :  2.990   Max.   :0.840   Max.   :1.000   Max.   :32.97  
#>  NA's   :4448      NA's   :4448    NA's   :13266   NA's   :4448   
#> dimension(s):
#>      from  to offset delta refsys point values    
#> lon     1 180     -1     2     NA    NA   NULL [x]
#> lat     1  90    -90     2     NA    NA   NULL [y]
#> zlev    1   1     NA    NA     NA    NA      0    
#> time    1   1     NA    NA     NA    NA   1460    

read_ncdf(f, var = c("anom"))

#> stars object with 4 dimensions and 1 attribute
#> attribute(s):
#>      anom         
#>  Min.   :-10.160  
#>  1st Qu.: -0.580  
#>  Median : -0.080  
#>  Mean   : -0.186  
#>  3rd Qu.:  0.210  
#>  Max.   :  2.990  
#>  NA's   :4448     
#> dimension(s):
#>      from  to offset delta refsys point values    
#> lon     1 180     -1     2     NA    NA   NULL [x]
#> lat     1  90    -90     2     NA    NA   NULL [y]
#> zlev    1   1     NA    NA     NA    NA      0    
#> time    1   1     NA    NA     NA    NA   1460    

read_ncdf(f, ncsub = cbind(start = c(1, 1, 1, 1), count = c(10, 12, 1, 1)))

#> stars object with 4 dimensions and 4 attributes
#> attribute(s):
#>      anom                err             ice              sst        
#>  Min.   :-1.07000   Min.   :0.300   Min.   :0.0100   Min.   :-1.390  
#>  1st Qu.:-0.36250   1st Qu.:0.300   1st Qu.:0.1100   1st Qu.:-0.720  
#>  Median : 0.19500   Median :0.300   Median :0.1700   Median :-0.515  
#>  Mean   : 0.05867   Mean   :0.303   Mean   :0.2094   Mean   :-0.534  
#>  3rd Qu.: 0.55500   3rd Qu.:0.300   3rd Qu.:0.2550   3rd Qu.:-0.275  
#>  Max.   : 0.92000   Max.   :0.320   Max.   :0.5200   Max.   : 0.030  
#>  NA's   :90         NA's   :90      NA's   :104      NA's   :90      
#> dimension(s):
#>      from to offset delta refsys point values    
#> lon     1 10     -1     2     NA    NA   NULL [x]
#> lat     1 12    -90     2     NA    NA   NULL [y]
#> zlev    1  1     NA    NA     NA    NA      0    
#> time    1  1     NA    NA     NA    NA   1460