spatially or temporally aggregate stars object

spatially or temporally aggregate stars object, returning a data cube with lower spatial or temporal resolution

# S3 method for stars
aggregate(x, by, FUN, ..., drop = FALSE,
  join = st_intersects, as_points = any(st_dimension(by) == 2, na.rm =
  TRUE), rightmost.closed = FALSE, left.open = FALSE)

Arguments

x
object of class stars with information to be aggregated

by
object of class sf or sfc for spatial aggregation, for temporal aggregation a vector with time values (Date, POSIXct, or PCICt) that is interpreted as a sequence of left-closed, right-open time intervals or a string like "months", "5 days" or the like (see cut.POSIXt); if by is an object of class stars, it is converted to sfc by st_as_sfc(by, as_points = FALSE) thus ignoring its time component.

FUN
aggregation function, such as mean

...
arguments passed on to FUN, such as na.rm=TRUE

drop
logical; ignored

join
join function to find matches of x to by

as_points
see st_as_sf: shall raster pixels be taken as points, or small square polygons?

rightmost.closed
see findInterval

left.open
logical; used for time intervals, see findInterval and cut.POSIXt