Read or write data using GDAL's multidimensional array API

Read or write data using GDAL's multidimensional array API

read_mdim(
  filename,
  variable = character(0),
  ...,
  options = character(0),
  raster = NULL,
  offset = integer(0),
  count = integer(0),
  step = integer(0),
  proxy = FALSE,
  debug = FALSE
)

write_mdim(
  x,
  filename,
  driver = detect.driver(filename),
  ...,
  root_group_options = character(0),
  options = character(0)
)

Arguments

filename

name of the source or destination file or data source

variable

name of the array to be read

...

ignored

options

character; driver specific options regarding the opening (read_mdim) or creation (write_mdim) of the dataset

raster

names of the raster variables (default: first two dimensions)

offset

integer; offset for each dimension (pixels) of sub-array to read (default: 0,0,0,...) (requires sf >= 1.0-9)

count

integer; size for each dimension (pixels) of sub-array to read (default: read all) (requires sf >= 1.0-9)

step

integer; step size for each dimension (pixels) of sub-aray to read (requires sf >= 1.0-9)

proxy

logical; return proxy object? (not functional yet)

debug

logical; print debug info?

x

stars object

driver

character; driver name

root_group_options

character; driver specific options regarding the creation of the root group

Details

it is assumed that the first two dimensions are easting / northing

Examples

set.seed(135)
m = matrix(runif(10), 2, 5)
names(dim(m)) = c("stations", "time")
times = as.Date("2022-05-01") + 1:5
pts = st_as_sfc(c("POINT(0 1)", "POINT(3 5)"))
s = st_as_stars(list(Precipitation = m)) |>
 st_set_dimensions(1, values = pts) |>
 st_set_dimensions(2, values = times)
nc = tempfile(fileext=".nc")
if (compareVersion(sf_extSoftVersion()["GDAL"], "3.4.0") > -1) {
  write_mdim(s, nc)
  # try ncdump on the generated file
  print(read_mdim(nc))
}