read raster/array dataset from file or connection

read raster/array dataset from file or connection

read_stars(.x, ..., options = character(0), driver = character(0),
  sub = TRUE, quiet = FALSE, NA_value = NA_real_,
  along = NA_integer_)

Arguments

.x
character vector with name(s) of file(s) or data source(s) to be read

...
ignored

options
character; opening options

driver
character; driver to use for opening file

sub
integer or logical; sub-datasets to be read

quiet
logical; print progress output?

NA_value
numeric value to be used for conversion into NA values; by default this is read from the input file

along
length-one character or integer, or list; determines how several arrays are combined, see Details.

Value

object of class stars

Details

In case .x contains multiple files, they will all be read and combined with c.stars. Along which dimension, or how should objects be merged? If along is set to NA it will merge arrays as new attributes if all objects have identical dimensions, or else try to merge along time if a dimension called time indicates different time stamps. A single name (or positive value) for along will merge along that dimension, or create a new one if it does not already exist. If the arrays should be arranged along one of more dimensions with values (e.g. time stamps), a named list can passed to along to specify them; see example.

Examples

tif = system.file("tif/L7_ETMs.tif", package = "stars")
(x1 = read_stars(tif))

#> stars object with 3 dimensions and 1 attribute
#> attribute(s):
#>   L7_ETMs.tif    
#>  Min.   :  1.00  
#>  1st Qu.: 54.00  
#>  Median : 69.00  
#>  Mean   : 68.91  
#>  3rd Qu.: 86.00  
#>  Max.   :255.00  
#> dimension(s):
#>      from  to  offset delta                       refsys point values
#> x       1 349  288776  28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> y       1 352 9120761 -28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> band    1   6      NA    NA                           NA    NA   NULL

(x2 = read_stars(c(tif, tif)))

#> stars object with 3 dimensions and 2 attributes
#> attribute(s):
#>   L7_ETMs.tif     L7_ETMs.tif.1   
#>  Min.   :  1.00   Min.   :  1.00  
#>  1st Qu.: 54.00   1st Qu.: 54.00  
#>  Median : 69.00   Median : 69.00  
#>  Mean   : 68.91   Mean   : 68.91  
#>  3rd Qu.: 86.00   3rd Qu.: 86.00  
#>  Max.   :255.00   Max.   :255.00  
#> dimension(s):
#>      from  to  offset delta                       refsys point values
#> x       1 349  288776  28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> y       1 352 9120761 -28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> band    1   6      NA    NA                           NA    NA   NULL

(x3 = read_stars(c(tif, tif), along = "band"))

#> stars object with 3 dimensions and 1 attribute
#> attribute(s), summary of first 1e+05 cells:
#>   L7_ETMs.tif    
#>  Min.   : 47.00  
#>  1st Qu.: 65.00  
#>  Median : 76.00  
#>  Mean   : 77.34  
#>  3rd Qu.: 87.00  
#>  Max.   :255.00  
#> dimension(s):
#>      from  to  offset delta                       refsys point values
#> x       1 349  288776  28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> y       1 352 9120761 -28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> band    1  12      NA    NA                           NA    NA   NULL

(x4 = read_stars(c(tif, tif), along = "new_dimensions")) # create 4-dimensional array

#> stars object with 4 dimensions and 1 attribute
#> attribute(s), summary of first 1e+05 cells:
#>   L7_ETMs.tif    
#>  Min.   : 47.00  
#>  1st Qu.: 65.00  
#>  Median : 76.00  
#>  Mean   : 77.34  
#>  3rd Qu.: 87.00  
#>  Max.   :255.00  
#> dimension(s):
#>                from  to  offset delta                       refsys point values
#> x                 1 349  288776  28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> y                 1 352 9120761 -28.5 +proj=utm +zone=25 +south... FALSE   NULL
#> band              1   6      NA    NA                           NA    NA   NULL
#> new_dimensions    1   2      NA    NA                           NA    NA   NULL

x1o = read_stars(tif, options = "OVERVIEW_LEVEL=1")
t1 = as.Date("2018-07-31")
# along is a named list indicating two dimensions:
read_stars(c(tif, tif, tif, tif), along = list(foo = c("bar1", "bar2"), time = c(t1, t1+2)))

#> stars object with 5 dimensions and 1 attribute
#> attribute(s), summary of first 1e+05 cells:
#>   L7_ETMs.tif    
#>  Min.   : 47.00  
#>  1st Qu.: 65.00  
#>  Median : 76.00  
#>  Mean   : 77.34  
#>  3rd Qu.: 87.00  
#>  Max.   :255.00  
#> dimension(s):
#>      from  to     offset  delta                       refsys point     values
#> x       1 349     288776   28.5 +proj=utm +zone=25 +south... FALSE       NULL
#> y       1 352    9120761  -28.5 +proj=utm +zone=25 +south... FALSE       NULL
#> band    1   6         NA     NA                           NA    NA       NULL
#> foo     1   2         NA     NA                           NA    NA bar1, bar2
#> time    1   2 2018-07-31 2 days                         Date    NA       NULL