convert objects into a stars object

convert objects into a stars object

st_as_stars(.x, ...)

# S3 method for list
st_as_stars(.x, ..., dimensions = NULL)

# S3 method for default
st_as_stars(.x = NULL, ..., raster = NULL)

# S3 method for stars
st_as_stars(.x, ..., curvilinear = NULL,
  crs = st_crs(4326))

# S3 method for bbox
st_as_stars(.x, ..., nx, ny, dx = dy, dy = dx,
  xlim = .x[c("xmin", "xmax")], ylim = .x[c("ymin", "ymax")],
  values = 0, n = 64800, pretty = FALSE, inside = FALSE)

# S3 method for sf
st_as_stars(.x, ..., name = attr(.x, "sf_column"))

# S3 method for Raster
st_as_stars(.x, ...)

# S3 method for stars_proxy
st_as_stars(.x, ..., downsample = 0,
  url = attr(.x, "url"), envir = parent.frame())

Arguments

.x
object to convert

...
in case .x is of class bbox, arguments passed on to pretty

dimensions
object of class dimensions

raster
character; the names of the dimensions that denote raster dimensions

curvilinear
only for creating curvilinear grids: named length 2 list holding longitude and latitude matrices; the names of this list should correspond to raster dimensions to be replaced

crs
object of class crs with the coordinate reference system of the values in curvilinear; see details

nx
integer; number of cells in x direction; see details

ny
integer; number of cells in y direction; see details

dx
numeric; cell size in x direction; see details

dy
numeric; cell size in y direction; see details

xlim
length 2 numeric vector with extent (min, max) in x direction

ylim
length 2 numeric vector with extent (min, max) in y direction

values
value(s) to populate the raster values with

n
the (approximate) target number of grid cells

pretty
logical; should cell coordinates have pretty values?

inside
logical; should all cells entirely fall inside the bbox, potentially not covering it completely?

name
character; name for the geometry dimensions

downsample
integer: if larger than 0, downsample with this rate (number of pixels to skip in every row/column); if length 2, specifies downsampling rate in x and y.

url
character; URL of the stars endpoint where the data reside

envir
environment to resolve objects in

Details

if curvilinear is a stars object with longitude and latitude values, its coordinate reference system is typically not that of the latitude and longitude values.

For the bbox method: if pretty is TRUE, raster cells may extend the coordinate range of .x on all sides. If in addition to nx and ny, dx and dy are also missing, these are set to a single value computed as sqrt(diff(xlim)*diff(ylim)/n). If nx and ny are missing, they are computed as the ceiling of the ratio of the (x or y) range divided by (dx or dy), unless inside is TRUE, in which ceiling is replaced by floor. Postive dy will be made negative. Further named arguments (...) are passed on to pretty.