Extract cell values at point locations

Extract cell values at point locations

st_extract(x, ...)

# S3 method for stars
st_extract(
  x,
  at,
  ...,
  bilinear = FALSE,
  time_column = attr(at, "time_column") %||% attr(at, "time_col"),
  interpolate_time = bilinear,
  FUN = mean
)

Arguments

x

object of class stars or stars_proxy

...

passed on to aggregate.stars when geometries are not exclusively POINT geometries

at

object of class sf or sfc with geometries, or two-column matrix with points in rows, indicating where to extract x

bilinear

logical; use bilinear interpolation rather than nearest neighbour?

time_column

character or integer; name or index of a column with time or date values that will be matched to values of the dimension "time" in x, after which this dimension is reduced. This is useful to extract data cube values along a trajectory; see https://github.com/r-spatial/stars/issues/352 .

interpolate_time

logical; should time be interpolated? if FALSE, time instances are matched using the coinciding or the last preceding time in the data cube.

FUN

function used to aggregate pixel values when geometries of at intersect with more than one pixel

Value

if at is of class matrix, a matrix with extracted values is returned; otherwise: if x has more dimensions than only x and y (raster), an object of class stars with POINT geometries replacing x and y raster dimensions, if this is not the case, an object of sf with extracted values.

Details

points outside the raster are returned as NA values. For large sets of points for which extraction is needed, passing a matrix as to at may be much faster than passing an sf or sfc object.

Examples

tif = system.file("tif/L7_ETMs.tif", package = "stars")
r = read_stars(tif)
pnt = st_sample(st_as_sfc(st_bbox(r)), 10)
st_extract(r, pnt)
#> stars object with 2 dimensions and 1 attribute
#> attribute(s):
#>              Min. 1st Qu. Median     Mean 3rd Qu. Max.
#> L7_ETMs.tif    29   52.75     66 69.66667   82.25  133
#> dimension(s):
#>          from to offset delta                     refsys point
#> geometry    1 10     NA    NA SIRGAS 2000 / UTM zone 25S  TRUE
#> band        1  6     NA    NA                         NA    NA
#>                                                       values
#> geometry POINT (290059.7 9111941),...,POINT (290848 9118009)
#> band                                                    NULL
st_extract(r, pnt) %>% st_as_sf()
#> Simple feature collection with 10 features and 6 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 288855.7 ymin: 9111941 xmax: 294806.9 ymax: 9118640
#> Projected CRS: SIRGAS 2000 / UTM zone 25S
#>    L7_ETMs.tif.V1 L7_ETMs.tif.V2 L7_ETMs.tif.V3 L7_ETMs.tif.V4 L7_ETMs.tif.V5
#> 1              81             68             84             54            133
#> 2              62             49             41             73             62
#> 3              66             53             49             54             78
#> 4              66             51             46             69             75
#> 5              84             72             82             65            115
#> 6              65             51             45             65             95
#> 7              74             60             67             48            113
#> 8              62             47             35             88             67
#> 9              88             79             88             73            133
#> 10             64             53             47             69             83
#>    L7_ETMs.tif.V6                 geometry
#> 1             117 POINT (290059.7 9111941)
#> 2              32 POINT (292752.2 9118640)
#> 3              56 POINT (289804.1 9115143)
#> 4              49 POINT (291456.8 9117264)
#> 5              90 POINT (293495.3 9114683)
#> 6              65 POINT (290272.7 9116389)
#> 7              94 POINT (288855.7 9113503)
#> 8              29 POINT (293135.6 9117098)
#> 9             105 POINT (294806.9 9117251)
#> 10             52   POINT (290848 9118009)
st_extract(r[,,,1], pnt)
#> Simple feature collection with 10 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 288855.7 ymin: 9111941 xmax: 294806.9 ymax: 9118640
#> Projected CRS: SIRGAS 2000 / UTM zone 25S
#>    L7_ETMs.tif                 geometry
#> 1           81 POINT (290059.7 9111941)
#> 2           62 POINT (292752.2 9118640)
#> 3           66 POINT (289804.1 9115143)
#> 4           66 POINT (291456.8 9117264)
#> 5           84 POINT (293495.3 9114683)
#> 6           65 POINT (290272.7 9116389)
#> 7           74 POINT (288855.7 9113503)
#> 8           62 POINT (293135.6 9117098)
#> 9           88 POINT (294806.9 9117251)
#> 10          64   POINT (290848 9118009)
st_extract(r, st_coordinates(pnt)) # "at" is a matrix: return a matrix
#>       [,1] [,2] [,3] [,4] [,5] [,6]
#>  [1,]   81   68   84   54  133  117
#>  [2,]   62   49   41   73   62   32
#>  [3,]   66   53   49   54   78   56
#>  [4,]   66   51   46   69   75   49
#>  [5,]   84   72   82   65  115   90
#>  [6,]   65   51   45   65   95   65
#>  [7,]   74   60   67   48  113   94
#>  [8,]   62   47   35   88   67   29
#>  [9,]   88   79   88   73  133  105
#> [10,]   64   53   47   69   83   52