Extract cell values at point locations

Extract cell values at point locations

st_extract(x, ...)

# S3 method for stars
st_extract(
  x,
  pts,
  ...,
  bilinear = FALSE,
  time_column = attr(pts, "time_column") %||% attr(pts, "time_col"),
  interpolate_time = bilinear
)

Arguments

x
object of class stars or stars_proxy

...
ignored

pts
object of class sf or sfc with POINT geometries

bilinear
logical; use bilinear interpolation rather than nearest neighbour?

time_column
character or integer; name or index of a column with time or date values that will be matched to values of the dimension "time" in x, after which this dimension is reduced. This is useful to extract data cube values along a trajectory; see https://github.com/r-spatial/stars/issues/352 .

interpolate_time
logical; should time be interpolated? if FALSE, time instances are matched using the coinciding or the last preceding time in the data cube.

Value

if x has more dimensions than only x and y (raster), an object of class stars with POINT geometries replacing x and y raster dimensions; otherwise an object of sf with extracted values.

Details

points outside the raster are returned as NA values.

Examples

tif = system.file("tif/L7_ETMs.tif", package = "stars")
r = read_stars(tif)
pnt = st_sample(st_as_sfc(st_bbox(r)), 10)
st_extract(r, pnt)

#> stars object with 2 dimensions and 1 attribute
#> attribute(s):
#>   L7_ETMs.tif    
#>  Min.   : 12.00  
#>  1st Qu.: 41.75  
#>  Median : 59.50  
#>  Mean   : 59.45  
#>  3rd Qu.: 76.50  
#>  Max.   :133.00  
#> dimension(s):
#>          from to offset delta                       refsys point
#> geometry    1 10     NA    NA UTM Zone 25, Southern Hem...  TRUE
#> band        1  6     NA    NA                           NA    NA
#>                                                         values
#> geometry POINT (289579.4 9119503),...,POINT (296460.1 9120515)
#> band                                                      NULL

st_extract(r, pnt) %>% st_as_sf()

#> Simple feature collection with 10 features and 6 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 288849.8 ymin: 9111072 xmax: 297074.9 ymax: 9120515
#> projected CRS:  UTM Zone 25, Southern Hemisphere
#>    L7_ETMs.tif.V1 L7_ETMs.tif.V2 L7_ETMs.tif.V3 L7_ETMs.tif.V4 L7_ETMs.tif.V5
#> 1              57             40             30             53             42
#> 2              97             93             78             15             14
#> 3              85             76             55             12             14
#> 4              72             56             59             47             91
#> 5              65             53             45             76             72
#> 6              80             65             71             53            120
#> 7              92             80             90             62            133
#> 8              81             66             64             50             82
#> 9              60             43             32             70             49
#> 10             69             50             48             41             36
#>    L7_ETMs.tif.V6                 geometry
#> 1              24 POINT (289579.4 9119503)
#> 2              13 POINT (297074.9 9112484)
#> 3              14 POINT (294751.7 9111072)
#> 4              72 POINT (290339.9 9113943)
#> 5              40 POINT (288849.8 9114765)
#> 6             100 POINT (293415.2 9112692)
#> 7             113 POINT (293727.4 9114777)
#> 8              64 POINT (291658.4 9111367)
#> 9              25 POINT (296065.9 9114628)
#> 10             18 POINT (296460.1 9120515)

st_extract(r[,,,1], pnt)

#> Simple feature collection with 10 features and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 288849.8 ymin: 9111072 xmax: 297074.9 ymax: 9120515
#> projected CRS:  UTM Zone 25, Southern Hemisphere
#>    L7_ETMs.tif                 geometry
#> 1           57 POINT (289579.4 9119503)
#> 2           97 POINT (297074.9 9112484)
#> 3           85 POINT (294751.7 9111072)
#> 4           72 POINT (290339.9 9113943)
#> 5           65 POINT (288849.8 9114765)
#> 6           80 POINT (293415.2 9112692)
#> 7           92 POINT (293727.4 9114777)
#> 8           81 POINT (291658.4 9111367)
#> 9           60 POINT (296065.9 9114628)
#> 10          69 POINT (296460.1 9120515)