Extract cell values at point locations

Extract cell values at point locations

st_extract(x, ...)

# S3 method for stars
st_extract(
  x,
  at,
  ...,
  bilinear = FALSE,
  time_column = attr(at, "time_column") %||% attr(at, "time_col"),
  interpolate_time = bilinear,
  FUN = mean
)

Arguments

x

object of class stars or stars_proxy

...

passed on to aggregate.stars when geometries are not exclusively POINT geometries

at

object of class sf or sfc with geometries, or two-column matrix with points in rows, indicating where to extract x

bilinear

logical; use bilinear interpolation rather than nearest neighbour?

time_column

character or integer; name or index of a column with time or date values that will be matched to values of the dimension "time" in x, after which this dimension is reduced. This is useful to extract data cube values along a trajectory; see https://github.com/r-spatial/stars/issues/352 .

interpolate_time

logical; should time be interpolated? if FALSE, time instances are matched using the coinciding or the last preceding time in the data cube.

FUN

function used to aggregate pixel values when geometries of at intersect with more than one pixel

Value

if at is of class matrix, a matrix with extracted values is returned; otherwise: if x has more dimensions than only x and y (raster), an object of class stars with POINT geometries replacing x and y raster dimensions, if this is not the case, an object of sf with extracted values.

Details

points outside the raster are returned as NA values. For large sets of points for which extraction is needed, passing a matrix as to at may be much faster than passing an sf or sfc object.

Examples

tif = system.file("tif/L7_ETMs.tif", package = "stars")
r = read_stars(tif)
pnt = st_sample(st_as_sfc(st_bbox(r)), 10)
st_extract(r, pnt)
#> stars object with 2 dimensions and 1 attribute
#> attribute(s):
#>              Min. 1st Qu. Median Mean 3rd Qu. Max.
#> L7_ETMs.tif    12      36     62 56.6   76.25  116
#> dimension(s):
#>          from to offset delta                     refsys point
#> geometry    1 10     NA    NA SIRGAS 2000 / UTM zone 25S  TRUE
#> band        1  6     NA    NA                         NA    NA
#>                                                         values
#> geometry POINT (296227.7 9114746),...,POINT (298229.2 9117908)
#> band                                                      NULL
st_extract(r, pnt) %>% st_as_sf()
#> Simple feature collection with 10 features and 6 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 291639.2 ymin: 9110771 xmax: 298626.1 ymax: 9120373
#> Projected CRS: SIRGAS 2000 / UTM zone 25S
#>    L7_ETMs.tif.V1 L7_ETMs.tif.V2 L7_ETMs.tif.V3 L7_ETMs.tif.V4 L7_ETMs.tif.V5
#> 1              62             46             36             84             57
#> 2              71             56             44             81             64
#> 3              79             66             67             60             94
#> 4              72             62             67             70            116
#> 5              89             81             60             13             12
#> 6              58             45             31             75             65
#> 7              87             79             52             12             12
#> 8              64             51             37             85             63
#> 9              96             88             71             14             15
#> 10             87             77             76             40             53
#>    L7_ETMs.tif.V6                 geometry
#> 1              26 POINT (296227.7 9114746)
#> 2              34 POINT (291639.2 9114820)
#> 3              69 POINT (293105.6 9120145)
#> 4              90 POINT (295713.4 9118887)
#> 5              12 POINT (295023.3 9110771)
#> 6              31 POINT (292143.3 9120373)
#> 7              14 POINT (296323.1 9111887)
#> 8              28 POINT (294762.9 9114801)
#> 9              14 POINT (298626.1 9118552)
#> 10             36 POINT (298229.2 9117908)
st_extract(r[,,,1], pnt)
#> Simple feature collection with 10 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 291639.2 ymin: 9110771 xmax: 298626.1 ymax: 9120373
#> Projected CRS: SIRGAS 2000 / UTM zone 25S
#>    L7_ETMs.tif                 geometry
#> 1           62 POINT (296227.7 9114746)
#> 2           71 POINT (291639.2 9114820)
#> 3           79 POINT (293105.6 9120145)
#> 4           72 POINT (295713.4 9118887)
#> 5           89 POINT (295023.3 9110771)
#> 6           58 POINT (292143.3 9120373)
#> 7           87 POINT (296323.1 9111887)
#> 8           64 POINT (294762.9 9114801)
#> 9           96 POINT (298626.1 9118552)
#> 10          87 POINT (298229.2 9117908)
st_extract(r, st_coordinates(pnt)) # "at" is a matrix: return a matrix
#>       [,1] [,2] [,3] [,4] [,5] [,6]
#>  [1,]   62   46   36   84   57   26
#>  [2,]   71   56   44   81   64   34
#>  [3,]   79   66   67   60   94   69
#>  [4,]   72   62   67   70  116   90
#>  [5,]   89   81   60   13   12   12
#>  [6,]   58   45   31   75   65   31
#>  [7,]   87   79   52   12   12   14
#>  [8,]   64   51   37   85   63   28
#>  [9,]   96   88   71   14   15   14
#> [10,]   87   77   76   40   53   36